+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `config` command to get and set configuration options.
-  - `-l, --list` flag to list all available configuration keys.
+- `config list|ls` command to list configuration options.
 - `endpoint list|ls` command to print the endpoints list as a table.
   - `--format` flag to format output using a Go template.
 - `stack list|ls` command to print the stacks list as a table.